Move ClockSkewHook down to the b2http module, and enable it by default.
