+++ test/test_b2http.py
@@ -8,14 +8,15 @@
 #
 ######################################################################
 
+import datetime
 import requests
 import six
 import socket
 import sys
 
 from .test_base import TestBase
-from b2.b2http import _translate_and_retry, _translate_errors, B2Http
-from b2.exception import BadJson, BrokenPipe, B2ConnectionError, ServiceError, UnknownError, UnknownHost
+from b2.b2http import _translate_and_retry, _translate_errors, B2Http, ClockSkewHook
+from b2.exception import BadDateFormat, BadJson, BrokenPipe, B2ConnectionError, ClockSkew, ServiceError, UnknownError, UnknownHost
 from b2.version import USER_AGENT
 
 if sys.version_info < (3, 3):
@@ -158,7 +159,7 @@ class TestB2Http(TestBase):
 
         requests = MagicMock()
         requests.Session.return_value = self.session
-        self.b2_http = B2Http(requests)
+        self.b2_http = B2Http(requests, install_clock_skew_hook=False)
 
     def test_post_json_return_json(self):
         self.session.post.return_value = self.response
@@ -195,3 +196,40 @@ class TestB2Http(TestBase):
             self.assertTrue(self.response is r)  # no assertIs until 2.7
         self.session.get.assert_called_with(self.URL, headers=self.EXPECTED_HEADERS, stream=True)
         self.response.close.assert_called_with()
+
+
+class TestClockSkewHook(TestBase):
+    def test_bad_format(self):
+        response = MagicMock()
+        response.headers = {'Date': 'bad format'}
+        with self.assertRaises(BadDateFormat):
+            ClockSkewHook().post_request('POST', 'http://foo.com', {}, response)
+
+    def test_bad_month(self):
+        response = MagicMock()
+        response.headers = {'Date': 'Fri, 16 XXX 2016 20:52:30 GMT'}
+        with self.assertRaises(BadDateFormat):
+            ClockSkewHook().post_request('POST', 'http://foo.com', {}, response)
+
+    def test_no_skew(self):
+        now = datetime.datetime.utcnow()
+        now_str = now.strftime('%a, %d %b %Y %H:%M:%S GMT')
+        response = MagicMock()
+        response.headers = {'Date': now_str}
+        ClockSkewHook().post_request('POST', 'http://foo.com', {}, response)
+
+    def test_positive_skew(self):
+        now = datetime.datetime.utcnow() + datetime.timedelta(minutes=11)
+        now_str = now.strftime('%a, %d %b %Y %H:%M:%S GMT')
+        response = MagicMock()
+        response.headers = {'Date': now_str}
+        with self.assertRaises(ClockSkew):
+            ClockSkewHook().post_request('POST', 'http://foo.com', {}, response)
+
+    def test_negative_skew(self):
+        now = datetime.datetime.utcnow() + datetime.timedelta(minutes=-11)
+        now_str = now.strftime('%a, %d %b %Y %H:%M:%S GMT')
+        response = MagicMock()
+        response.headers = {'Date': now_str}
+        with self.assertRaises(ClockSkew):
+            ClockSkewHook().post_request('POST', 'http://foo.com', {}, response)
